Show tests work out of crate.
